refactor(schemas): move audit tables to schemas/audit/
Per-dialect Drizzle table definitions for auditLog and activityLog now live
in schemas/audit/{postgres,mysql,sqlite}.ts with a dialect-aware barrel
exposing auditTables(dialect).

database/schema/<dialect>.ts re-exports from the new location so existing
imports and the activityLogRelations() block keep compiling. Relations
themselves stay in database/schema/<dialect>.ts during Plan A and
consolidate in Task 17.
