+++ packages/nextly/src/database/schema/mysql.ts
@@ -59,28 +59,11 @@ export const contentSchemaEvents = mysqlTable(
   ]
 );
 
-// Append-only by application convention — operators should revoke
-// UPDATE / DELETE GRANTs on this table in production for stricter
-// integrity.
-export const auditLog = mysqlTable(
-  "audit_log",
-  {
-    id: varchar("id", { length: 191 }).primaryKey(),
-    kind: varchar("kind", { length: 64 }).notNull(),
-    actorUserId: varchar("actor_user_id", { length: 191 }),
-    targetUserId: varchar("target_user_id", { length: 191 }),
-    ipAddress: varchar("ip_address", { length: 45 }),
-    userAgent: text("user_agent"),
-    metadata: json("metadata"),
-    createdAt: datetime("created_at").notNull().default(new Date()),
-  },
-  t => [
-    index("audit_log_kind_idx").on(t.kind),
-    index("audit_log_actor_user_id_idx").on(t.actorUserId),
-    index("audit_log_target_user_id_idx").on(t.targetUserId),
-    index("audit_log_created_at_idx").on(t.createdAt),
-  ]
-);
+// Audit tables — moved to schemas/audit/mysql.ts (Plan A Task 9).
+// Re-exported here so existing consumers and relations() blocks below keep working
+// until Task 16 sweeps imports to @nextly/schemas.
+export { auditLog, activityLog } from "../../schemas/audit/mysql";
+import { auditLog, activityLog } from "../../schemas/audit/mysql";
 
 // -----------------------------
 // RBAC tables (roles/permissions) — moved to schemas/rbac/mysql.ts (Plan A Task 7).
@@ -173,36 +156,8 @@ export const apiKeys = mysqlTable(
 // Activity Log
 // -----------------------------
 
-/**
- * Activity log table for recording user actions across all collections (MySQL).
- *
- * See postgres.ts for detailed documentation.
- * Main differences:
- * - Uses varchar(191) for string IDs (MySQL utf8mb4 index length limit)
- * - Uses datetime for timestamps
- */
-export const activityLog = mysqlTable(
-  "activity_log",
-  {
-    id: varchar("id", { length: 191 }).primaryKey(),
-    userId: varchar("user_id", { length: 191 })
-      .notNull()
-      .references(() => users.id, { onDelete: "cascade" }),
-    userName: varchar("user_name", { length: 255 }).notNull(),
-    userEmail: varchar("user_email", { length: 255 }).notNull(),
-    action: varchar("action", { length: 10 }).notNull(), // 'create' | 'update' | 'delete'
-    collection: varchar("collection", { length: 255 }).notNull(),
-    entryId: varchar("entry_id", { length: 191 }),
-    entryTitle: text("entry_title"),
-    metadata: text("metadata"), // JSON string for additional context
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  t => [
-    index("idx_activity_log_created_at").on(t.createdAt),
-    index("idx_activity_log_collection").on(t.collection, t.createdAt),
-    index("idx_activity_log_user_id").on(t.userId, t.createdAt),
-  ]
-);
+// activityLog moved to schemas/audit/mysql.ts (Plan A Task 9) and re-exported
+// with auditLog above.
 
 // -----------------------------
 // Relations
