+++ packages/nextly/src/schemas/index.ts
@@ -85,6 +85,9 @@ export {
 
 // Plan A Task 8 — media tables. PG re-exports for direct-query callers.
 export { media, mediaFolders, imageSizes } from "./media/postgres";
+
+// Plan A Task 9 — audit tables. PG re-exports for direct-query callers.
+export { auditLog, activityLog } from "./audit/postgres";
 export * from "./dynamic-collections"; // dialect-aware barrel — kept; unchanged
 export * from "./dynamic-components"; // kept; unchanged
 export * from "./migrations"; // dropped in Plan B; kept here in Plan A
